requirements.txt
Flask>=3.0
Pillow>=10.0
pytesseract>=0.3.10
